Enum stackedconfig::Lookup

pub enum Lookup {
    Missing,
    Found(Value),
}

Return value from looking up a path from a ConfigStack

conf.get("/foo/bar")      // Lookup::Found(1)
conf.get("/foo/bar/baz")  // Lookup::Missing

Variants
Missing

Indicates that the path did not resolve to a value

Found(Value)

Contains the serde_json::Value found from a lookup